requirements.txt
requests
pytest
jsonschema
python-dotenv
pytest-html
allure-pytest
requests-mock
pandas
numpy
pytest-order
pytest-dependency
json-server
Selenium
pytest
pytest-html
allure-pytest

